requirements.txt
PyPDF2==3.0.1
python_docx==1.1.0
pyttsx3==2.90
SpeechRecognition==3.10.3
streamlit==1.29.0
